% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_calculation.R
\name{fill_na_mean}
\alias{fill_na_mean}
\title{Replace missing values with the mean}
\usage{
fill_na_mean(numeric.vector)
}
\arguments{
\item{numeric.vector}{A vector with at least one numeric value.}
}
\value{
The input whose NAs are replaced by the input's mean.
}
\description{
Replace missing values with the mean
}
\examples{
fill_na_mean(c(1, 2, 3, NA, 5, 3, NA, 6))
}
